requirements2.txt
aiofiles==23.2.1
aiohttp==3.9.1
aiosignal==1.3.1
aiostream==0.5.2
alembic==1.13.0
altair==5.2.0
aniso8601==9.0.1
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.3
atlassian-python-api==3.41.4
atpublic==4.0
attrs==23.1.0
backoff==2.2.1
bcrypt==4.1.1
beautifulsoup4==4.12.2
bidict==0.22.1
bitarray==2.9.2
blinker==1.7.0
boilerpy3==1.0.7
Brotli==1.1.0
cachelib==0.10.2
cachetools==5.3.2
cattrs==23.2.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.15
ci-info==0.3.0
click==8.1.7
cloudpickle==2.2.1
coloredlogs==15.0.1
configobj==5.0.8
configparser==6.0.0
contourpy==1.2.0
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.12.1
dataclasses-json==0.6.3
Deprecated==1.2.14
distro==1.9.0
dnspython==2.4.2
docopt==0.6.2
effdet==0.4.1
emoji==2.9.0
environs==9.5.0
etelemetry==0.3.1
Events==0.5
exceptiongroup==1.2.0
faiss-cpu==1.7.4
farm-haystack==1.23.0
fastapi==0.105.0
filelock==3.13.1
filetype==1.2.0
fitz==0.0.1.dev2
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Login==0.6.3
Flask-Mail==0.9.1
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-Session==0.5.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-Uploads==0.2.1
flatbuffers==23.5.26
fonttools==4.47.0
frontend==0.0.3
frozenlist==1.4.0
fsspec==2023.12.1
future==0.18.3
gitdb==4.0.11
GitPython==3.1.40
google-auth==2.25.2
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.58.0
gunicorn==21.2.0
h11==0.14.0
html2text==2020.1.16
html5lib==1.1
httpcore==1.0.2
httplib2==0.22.0
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.20.1
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib-resources==6.1.1
inflect==7.0.0
iopath==0.1.10
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.329
langchain-core==0.0.12
langdetect==1.0.9
langsmith==0.0.69
layoutparser==0.3.4
lazy-imports==0.3.1
llama-hub==0.0.44
llama-index==0.9.7
looseversion==1.3.0
lxml==4.9.3
Mako==1.3.0
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
mdurl==0.1.2
minio==7.2.3
mmh3==4.0.1
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
networkx==3.2.1
nibabel==5.1.0
nipype==1.8.6
nltk==3.8.1
num2words==0.5.13
numexpr==2.8.8
numpy==1.26.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==0.28.0
opencv-python==4.8.1.78
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
overrides==7.4.0
packaging==23.2
pandas==2.1.4
pathlib==1.0.1
pdf2image==1.16.3
pdfkit==1.0.0
pdfminer.six==20221105
pdfplumber==0.10.3
pdfreader==0.1.12
pikepdf==8.10.1
Pillow==10.1.0
platformdirs==3.11.0
portalocker==2.8.2
posthog==3.1.0
prompthub-py==4.0.0
protobuf==4.21.12
prov==2.0.0
psutil==5.9.7
pulsar-client==3.3.0
pyarrow==14.0.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.1
pydantic==1.10.13
pydantic_core==2.14.5
pydeck==0.8.1b0
pydot==1.4.2
pydyf==0.8.0
Pygments==2.17.2
PyJWT==2.8.0
pymilvus==2.3.5
pymongo==4.6.1
PyMuPDF==1.23.8
PyMuPDFb==1.23.7
pyOpenSSL==23.3.0
pyparsing==3.1.1
pypdf==3.17.3
PyPDF2==3.0.1
pypdfium2==4.25.0
pyphen==0.14.0
PyPika==0.48.9
pytesseract==0.3.10
python-dateutil==2.8.2
python-dotenv==1.0.0
python-engineio==4.8.2
python-iso639==2023.12.11
python-magic==0.4.27
python-multipart==0.0.6
python-socketio==5.11.0
pytils==0.4.1
pytz==2023.3.post1
pyxnat==1.6
PyYAML==6.0.1
quantulum3==0.9.0
rank-bm25==0.2.2
rapidfuzz==3.5.2
rdflib==7.0.0
referencing==0.32.0
regex==2023.10.3
reportlab==4.0.8
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.7.0
rpds-py==0.13.2
rsa==4.9
safetensors==0.4.1
scikit-learn==1.3.2
scipy==1.11.4
simple-websocket==1.0.0
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
snowflake==0.5.1
snowflake-connector-python==3.6.0
snowflake-snowpark-python==1.11.1
snowflake._legacy==0.4.0
snowflake.core==0.5.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.23
sseclient-py==1.8.0
starlette==0.27.0
streamlit==1.29.0
streamlit-pills==0.3.0
stripe==7.12.0
summarizer==0.0.7
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.2
timm==0.9.12
tinycss2==1.2.1
tokenizers==0.15.0
toml==0.10.2
tomlkit==0.12.3
tools==0.1.9
toolz==0.12.0
torch==2.1.1
torchvision==0.16.1
tornado==6.4
tqdm==4.66.1
traits==6.3.2
transformers==4.35.2
triton==2.1.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
ujson==5.9.0
unstructured==0.11.6
unstructured-client==0.15.0
unstructured-inference==0.7.18
unstructured.pytesseract==0.3.12
url-normalize==1.4.3
urllib3==1.26.18
uvicorn==0.24.0.post1
uvloop==0.19.0
validators==0.22.0
watchdog==3.0.0
watchfiles==0.21.0
weasyprint==60.2
webencodings==0.5.1
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.1
wkhtmltopdf==0.2
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
zopfli==0.2.3

